g_ads/requirements.txt
argcomplete==1.12.3
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.10
colorlog==6.6.0
distlib==0.3.4
facebook-sdk==3.1.0
filelock==3.4.2
fsspec==2022.1.0
google-ads==14.1.0
google-api-core==2.5.0
google-auth==2.6.0
google-auth-oauthlib==0.4.6
googleapis-common-protos==1.54.0
grpcio==1.43.0
idna==3.3
nox==2022.1.7
numpy==1.22.1
oauthlib==3.2.0
packaging==21.3
pandas==1.3.5
platformdirs==2.5.0
proto-plus==1.18.1
protobuf==3.17.3
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.7
python-dateutil==2.8.2
pytz==2021.3
PyYAML==5.4.1
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
six==1.16.0
urllib3==1.26.8
virtualenv==20.13.1
